Adding threading for when not running against PouchDB.
